+++ src/utils/buildSchema.ts
@@ -25,8 +25,8 @@ function getEmitSchemaDefinitionFileOptions(buildSchemaOptions: BuildSchemaOptio
       typeof buildSchemaOptions.emitSchemaFile === "string"
         ? buildSchemaOptions.emitSchemaFile
         : typeof buildSchemaOptions.emitSchemaFile === "object"
-        ? buildSchemaOptions.emitSchemaFile.path || defaultSchemaFilePath
-        : defaultSchemaFilePath,
+          ? buildSchemaOptions.emitSchemaFile.path || defaultSchemaFilePath
+          : defaultSchemaFilePath,
     printSchemaOptions:
       typeof buildSchemaOptions.emitSchemaFile === "object"
         ? { ...defaultPrintSchemaOptions, ...buildSchemaOptions.emitSchemaFile }